% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{available_languages}
\alias{available_languages}
\title{Get Available Languages for BirdNET Model}
\usage{
available_languages(version)
}
\arguments{
\item{version}{character. The version of BirdNET to use (default is "v2.4", no other versions are currently supported).}
}
\value{
A sorted character vector containing the available language codes.
}
\description{
Retrieve the available languages supported by a specific version of BirdNET.
}
\examples{
\dontrun{available_languages("v2.4")}
}
