% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet-package.R
\docType{package}
\name{birdnet}
\alias{birdnet-package}
\alias{birdnet}
\alias{birdnetR}
\title{BirdNET for R}
\description{
\code{birdnetR} is a wrapper around the python package \code{birdnet}.
}
\keyword{internal}
