% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnetR-defunct.R
\name{birdnetR-defunct}
\alias{birdnetR-defunct}
\title{Defunct Functions in the birdnetR Package}
\description{
These functions are gone and no longer available.
}
\section{Defunct functions}{

\describe{
\item{\code{\link{install_birdnet}}}{This function is defunct. Python dependencies are managed using \code{reticulate::py_require()}}
}
}

\keyword{internal}
