% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnetR-deprecated.R
\name{birdnetR-deprecated}
\alias{birdnetR-deprecated}
\title{Deprecated Functions in the birdnetR Package}
\description{
These functions are deprecated and will be removed in future versions of the birdnetR package.
Please use the alternatives listed below.
}
\section{Deprecated functions}{

\describe{
\item{\code{\link{init_model}}}{This function is deprecated. Use \code{\link{birdnet_model_tflite}} instead.}
}
}

\keyword{internal}
