% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{.check_arrow}
\alias{.check_arrow}
\title{Check Arrow Package Availability}
\usage{
.check_arrow()
}
\value{
A named logical vector indicating availability in R and Python
}
\description{
This function checks if the Arrow package is installed and available in both R and Python.
}
\keyword{internal}
