% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{.check_birdnet_version}
\alias{.check_birdnet_version}
\title{Check the Installed birdnet Version}
\usage{
.check_birdnet_version()
}
\value{
None. This function is called for its side effect of stopping execution if the wrong version is installed.
}
\description{
This internal function checks if birdnet Python is installed and if the version matches the requirement.
If it is not available or if the versions do not match, issue a warning with instructions to update the package.
}
\keyword{internal}
