% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize BirdNET-R Package}
\usage{
.onLoad(libname, pkgname, ...)
}
\arguments{
\item{libname}{The name of the library currently being loaded.}

\item{pkgname}{The name of the package currently being loaded.}

\item{...}{Additional arguments passed to the function.}
}
\description{
This function is executed when the BirdNET-R package is loaded. It sets up the Python environment using the \code{reticulate} package, ensuring that the necessary Python dependencies are available. The function configures the Python virtual environment named \code{r-birdnet} and imports the required Python modules, including \code{birdnet.models} and \code{pathlib}.
}
