% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{.required_birdnet_version}
\alias{.required_birdnet_version}
\title{Get the Required BirdNET Version}
\usage{
.required_birdnet_version()
}
\value{
A string representing the required BirdNET version.
}
\description{
This internal function returns the required version of the BirdNET Python package.
Update this function with the new required version when the package is updated.
}
\examples{
.required_birdnet_version()
}
\keyword{internal}
