% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{.suggested_python_version}
\alias{.suggested_python_version}
\title{Get the Suggested Python Version}
\usage{
.suggested_python_version()
}
\value{
A string representing the suggested Python version.
}
\description{
This internal function returns the suggested version of Python to be used with the BirdNET package.
Update this function with the new suggested version when necessary.
}
\keyword{internal}
