% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{get_labels_path}
\alias{get_labels_path}
\title{Get Path to BirdNET Labels File for a Specified Language}
\usage{
get_labels_path(language)
}
\arguments{
\item{language}{A character string specifying the language code for which the labels path is requested.
The language must be one of the available languages supported by the BirdNET model.}
}
\value{
A character string representing the file path to the labels file for the specified language.
}
\description{
This function retrieves the file path to the BirdNET labels file on your system corresponding to a specified language.
This file contains all class labels supported by the BirdNET model.
}
\note{
The \code{language} parameter must be one of the available languages returned by \code{available_languages()}.
}
\examples{
get_labels_path("en_us")
}
\seealso{
\code{\link[=available_languages]{available_languages()}}
}
