% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{get_species_from_file}
\alias{get_species_from_file}
\title{Read species labels from a file}
\usage{
get_species_from_file(species_file)
}
\arguments{
\item{species_file}{Path to species file.}
}
\value{
A vector with class labels e.g. c("Cyanocitta cristata_Blue Jay", "Zenaida macroura_Mourning Dove")
}
\description{
This is a convenience function to read species labels from a file.
}
\examples{
# Read a custom species file
get_species_from_file(system.file("extdata", "species_list.txt", package = "birdnetR"))

# To access all class labels that are supported in your language,
# you can read in the respective label file
model <- birdnet_model_tflite(version = "v2.4", language = "en_us")
labels_path <- get_labels_path(model, "fr")
species_list <- get_species_from_file(labels_path)
head(species_list)
}
\seealso{
\code{\link[=available_languages]{available_languages()}} \code{\link[=get_labels_path]{get_labels_path()}}
}
