% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{init_model}
\alias{init_model}
\title{Initialize the BirdNET Model}
\usage{
init_model()
}
\value{
An instance of the BirdNET model.
}
\description{
This function initializes the BirdNET model (v2.4).
}
