% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{model_factory}
\alias{model_factory}
\title{Dynamically create a BirdNET model}
\usage{
model_factory(model_name, version, ...)
}
\arguments{
\item{model_name}{Character. The name of the model to create (e.g., "tflite", "protobuf").}

\item{version}{Character. The version of the model (e.g., "v2.4").}

\item{...}{Additional arguments passed to the Python model constructor (e.g., \code{tflite_num_threads}, \code{language}).}
}
\value{
A BirdNET model object of class \code{birdnet_model} and its subclasses (e.g., "tflite_v2.4").
}
\description{
This function dynamically creates a BirdNET model based on the provided model name and version. It retrieves
the appropriate Python model constructor from the module map, evaluates the constructor, and returns a wrapped
BirdNET model object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
py_birdnet_models <- reticulate::import("birdnet.models")
birdnet_model <-
  birdnetR:::model_factory(
    "tflite",
    "v2.4",
    tflite_num_threads = 2,
    language = "en_us"
  )
\dontshow{\}) # examplesIf}
}
\keyword{internal}
