% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{predict_species}
\alias{predict_species}
\title{Predict Species Within an Audio File}
\usage{
predict_species(
  model,
  audio_file = system.file("extdata", "soundscape.wav", package = "birdnetR"),
  min_confidence = 0.1,
  batch_size = 1L,
  chunk_overlap_s = 0,
  use_bandpass = TRUE,
  bandpass_fmin = 0L,
  bandpass_fmax = 15000L,
  apply_sigmoid = TRUE,
  sigmoid_sensitivity = 1,
  filter_species = NULL,
  keep_empty = TRUE
)
}
\arguments{
\item{model}{BirdNETModel. An instance of the BirdNET model returned by \code{\link[=init_model]{init_model()}}.}

\item{audio_file}{character. The path to the audio file.}

\item{min_confidence}{numeric. Minimum confidence threshold for predictions.}

\item{batch_size}{integer. Number of audio samples to process in a batch.}

\item{chunk_overlap_s}{numeric. Overlapping of chunks in seconds. Must be in the interval [0.0, 3.0\).}

\item{use_bandpass}{logical. Whether to apply a bandpass filter.}

\item{bandpass_fmin, bandpass_fmax}{numeric. Minimum/Maximum frequency for the bandpass filter (in Hz). Ignored if \code{use_bandpass} is False.}

\item{apply_sigmoid}{logical. Whether to apply a sigmoid function to the model output.}

\item{sigmoid_sensitivity}{numeric. Sensitivity parameter for the sigmoid function. Must be in the interval 0.5 - 1.5. Ignored if \code{apply_sigmoid} is False.}

\item{filter_species}{NULL, a character vector of length greater than 0 or a list where each element is a single non-empty character string. Used to filter the predictions. If NULL, no filtering is applied. See \code{\link[=get_species_from_file]{get_species_from_file()}} for more details.}

\item{keep_empty}{logical. Whether to include empty intervals in the output.}
}
\value{
A data frame with columns: \code{start}, \code{end}, \code{scientific_name}, \code{common_name}, and \code{confidence}.
Each row represents a single prediction.
}
\description{
This function predicts species within an audio file using the BirdNET model.
}
\details{
Applying a sigmoid activation function, (\code{apply_sigmoid=True}) scales the unbound class output of the linear classifier ("logit score") to the range \code{0-1}.
This confidence score is a unitless, numeric expression of BirdNET’s “confidence” in its prediction (but not the probability of species presence).
Sigmoid sensitivity < 1 leads to more higher and lower scoring predictions and a value > 1 leads to more intermediate-scoring predictions.

For more information on BirdNET confidence scores, the sigmoid activation function and a suggested workflow on how to convert confidence scores to probabilities, see Wood & Kahl, 2024
}
\references{
Wood, C. M., & Kahl, S. (2024). Guidelines for appropriate use of BirdNET scores and other detector outputs. Journal of Ornithology. https://doi.org/10.1007/s10336-024-02144-5
}
\seealso{
\code{\link[=init_model]{init_model()}} \code{\link[=get_species_from_file]{get_species_from_file()}}
}
