% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predictions_list_element_to_df}
\alias{predictions_list_element_to_df}
\title{Convert a single prediction element to a data frame}
\usage{
predictions_list_element_to_df(x)
}
\arguments{
\item{x}{A single list element from the predictions list. It is expected to be a named list
with one or more elements where the names represent labels "scientificName_commonName" and
the values are confidence scores.}
}
\value{
A data frame with columns: \code{start}, \code{end}, \code{scientific_name}, \code{common_name}, and \code{confidence}.
}
\description{
This helper function takes a single list element from the predictions list and parses it into
a data frame format, extracting the time interval, scientific name, common name, and confidence level.
}
\keyword{internal}
