% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{available_languages}
\alias{available_languages}
\title{Get Available Languages for BirdNET Model}
\usage{
available_languages()
}
\value{
A sorted character vector containing the available language codes.
}
\description{
Retrieve the available languages supported by the BirdNET model.
}
\examples{
  available_languages()
}
