% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet-package.R
\docType{package}
\name{birdnetR}
\alias{birdnetR-package}
\alias{birdnetR}
\title{BirdNET for R}
\description{
\code{birdnetR} is a wrapper around the python package \code{birdnet}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://birdnet-team.github.io/birdnetR/}
}

}
\author{
\strong{Maintainer}: Stefan Kahl \email{stefan.kahl@cornell.edu}

Authors:
\itemize{
  \item BirdNET Team
}

}
\keyword{internal}
