% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{.check_birdnet_version}
\alias{.check_birdnet_version}
\title{Check the Installed BirdNET Version}
\usage{
.check_birdnet_version()
}
\value{
None. This function is called for its side effect of stopping execution if the wrong version is installed.
}
\description{
This internal function checks if the installed version of the BirdNET Python package matches the required version.
If the versions do not match, an error is raised with instructions to update the package.
}
\keyword{internal}
