% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_map.R
\name{evaluate_python_path}
\alias{evaluate_python_path}
\title{Evaluate a Python path string and return the corresponding Python object}
\usage{
evaluate_python_path(path_string)
}
\arguments{
\item{path_string}{Character. The string representing the Python path (e.g., "py_birdnet_models$v2m4$AudioModelV2M4TFLite").}
}
\value{
The evaluated Python object or value.
}
\description{
This function takes a string representing a Python path (e.g., from \code{get_model_from_module_map()})
and evaluates it to return the corresponding Python object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
py_birdnet_models <- reticulate::import("birdnet.models")
module_map <- create_module_map("v2.4", "py_birdnet_models")
model_string <- get_model_from_module_map(module_map, "tflite_v2.4")
model_object <- evaluate_python_path(model_string)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
