% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{get_top_prediction}
\alias{get_top_prediction}
\title{Get Top Prediction within a Time Interval}
\usage{
get_top_prediction(predictions, start_time, end_time)
}
\arguments{
\item{predictions}{A list of predictions from the BirdNET model.}

\item{start_time}{The start time of the interval.}

\item{end_time}{The end time of the interval.}
}
\value{
A list containing the top prediction and its confidence.
}
\description{
This function retrieves the most probable prediction within a specified time interval.
}
