% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{init_model}
\alias{init_model}
\title{Initialize the BirdNET Model}
\usage{
init_model(tflite_num_threads = NULL, language = "en_us")
}
\arguments{
\item{tflite_num_threads}{integer. The number of threads to use for TensorFlow Lite operations. If NULL (default), the default threading behavior will be used.
Will be coerced to an integer if possible.}

\item{language}{A character string specifying the language code to use for the model's text processing. The language must be one of the available languages supported by the BirdNET model.}
}
\value{
An instance of the BirdNET model.
}
\description{
This function initializes the BirdNET model (v2.4).
}
\note{
The \code{language} parameter must be one of the available languages returned by \code{available_languages()}.
}
\seealso{
\code{\link[=available_languages]{available_languages()}}
}
