% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_arrow}
\alias{install_arrow}
\title{Install Apache Arrow}
\usage{
install_arrow()
}
\value{
Invisible TRUE if successful, stops with error message if installation fails
}
\description{
This helper function installs Apache Arrow for both R and Python.
}
\examples{
\dontrun{install_arrow()}
}
