% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_birdnet}
\alias{install_birdnet}
\title{Install BirdNET and its dependencies (Defunct)}
\usage{
install_birdnet(...)
}
\arguments{
\item{...}{Arguments previously passed to this function (ignored)}
}
\description{
This function is defunct and should not be used.
}
\details{
The package functions and Python dependencies are now automatically handled
by the reticulate package via \code{reticulate::py_require()}.
}
\seealso{
\code{\link{Defunct}}
}
\keyword{defunct}
\keyword{internal}
