% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_birdnet}
\alias{install_birdnet}
\title{Install BirdNET and its dependencies}
\usage{
install_birdnet(
  ...,
  method = NULL,
  envname = "r-birdnet",
  new_env = identical(envname, "r-birdnet"),
  include_arrow = FALSE
)
}
\arguments{
\item{...}{Further arguments passed to \code{reticulate::py_install()}}

\item{method}{Installation method. Defaults to 'virtualenv' on macOS and Linux, and 'auto' on Windows. See \code{reticulate::py_install()} for more details.}

\item{envname}{Name of the virtual environment.}

\item{new_env}{If \code{TRUE}, any existing Python virtual environment specified by \code{envname} is deleted first.}

\item{include_arrow}{If \code{TRUE}, also installs the Apache Arrow for both R and Python for optimized data conversion. Arrow can be installed later using \code{install_arrow()}.}
}
\description{
\code{install_birdnet()} installs the python package \code{birdnet} from PyPi into a virtual environment.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
install_birdnet(method = "virtualenv", envname = "r-birdnet", include_arrow = TRUE)
\dontshow{\}) # examplesIf}
}
