% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_birdnet}
\alias{install_birdnet}
\title{Install BirdNET and its dependencies}
\usage{
install_birdnet(
  ...,
  envname = "r-birdnet",
  new_env = identical(envname, "r-birdnet")
)
}
\arguments{
\item{...}{Further arguments passed to \code{reticulate::py_install()}}

\item{envname}{Name of the virtual environment. Defaults to 'r-birdnet'.}

\item{new_env}{If \code{TRUE}, any existing Python virtual environment specified by \code{envname} is deleted first.}
}
\description{
\code{install_birdnet()} installs the python package \code{birdnet} from PyPi into a virtual environment.
}
