% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{install_birdnet}
\alias{install_birdnet}
\title{Install BirdNET}
\usage{
install_birdnet(..., envname = "r-birdnet")
}
\arguments{
\item{...}{Further arguments passed to reticulate::py_install()}

\item{envname}{Name of the virtual environment. Defaults to 'r-birdnet'.}
}
\description{
This function installs the python package \code{birdnet} from PyPi into a virtual environment.
}
