% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_load.R
\name{load_birdnet_model}
\alias{load_birdnet_model}
\alias{load_model}
\alias{load_custom}
\title{Load a BirdNET Model}
\usage{
load_model(
  type = "acoustic",
  version = "2.4",
  backend = "tf",
  library = NULL,
  precision = "fp32",
  language = "en_us"
)

load_custom(
  type = "acoustic",
  version = "2.4",
  backend = "tf",
  library = NULL,
  precision = "fp32",
  model = NULL,
  species_list = NULL,
  check_validity = TRUE,
  classifier_type = NULL,
  is_raven = NULL
)
}
\arguments{
\item{language}{character. Language code for the model to use e.g., "en_us".
Common species names are returned in the specified language if available. Use \code{\link[=available_languages]{available_languages()}} to see all available languages.}

\item{model}{If backend is "tf", path to custom model file. If backend is "pb", path to model directory.}

\item{species_list}{Path to the species list file.}

\item{check_validity}{Checks if the model is loadable by loading the model twice.}

\item{classifier_type}{Advanced option for custom TensorFlow models. Controls how the custom classifier head is interpreted.}

\item{is_raven}{Advanced option for custom protobuf models. Indicates whether the model uses the Raven protobuf layout.}
}
\value{
A BirdNET model object, which is an S3 object of class \code{birdnet_model} and specific subclasses (e.g., \code{birdnet_model_tflite}, \code{birdnet_model_v2_4}). This object is a list containing:
\describe{
\item{\code{py_model}}{The underlying Python BirdNET model object.}
\item{\code{model_type}}{The type of the model, either "acoustic" or "geo".}
\item{\code{model_version}}{The version string of the model (e.g., "v2.4").}
\item{\code{precision}}{The precision of the model, e.g., "int8", "fp16", "fp32".}
\item{...}{Additional elements specific to the model type:}
\itemize{
\item \strong{For pretrained acoustic and geo models}:
\itemize{
\item \code{lang}: The language code used (e.g., "en_us").
\item \code{library}: The library used for the model, if applicable.
}
\item \strong{For custom models}:
\itemize{
\item \code{model_path}: Path to custom model file (TensorFlow backend) or directory with model files (Protobuf backend).
\item \code{labels_path}: Path to the labels file.
}
}
\description{
Functions to load BirdNET models for sound identification or species prediction from location and time.
Models will be downloaded if not available locally.
\itemize{
\item \code{\link[=load_model]{load_model()}}: load a pre-trained BirdNET model or a geographic model.
\item \code{\link[=load_custom]{load_custom()}}: load a custom trained BirdNET model.
}
}
\details{
The argument \code{type} specifies the type of model to load:
\itemize{
\item \code{"acoustic"}: A model for species identification from audio recordings.
\item \code{"geo"}: A model for species prediction based on geographic location and time.
}

The argument \code{version} specifies the version of the model to load, e.g., \code{"2.4"}. Find all suported versions

The argument \code{backend} specifies the backend to use:
\itemize{
\item \code{"tf"}: TensorFlow backend, which supports different precisions and libraries but is CPU only.
\item \code{"pb"}: Protobuf backend, which can be run on a GPU. TODO describe how to insall GPU backends.
}

The argument \code{library} is only used for the TensorFlow backend and specifies the library to use:
\itemize{
\item \code{"litert"}: The LiteRT library for running TensorFlow Lite models.
\item \code{"tflite"}: The standard TensorFlow Lite library.
}

The argument \code{precision} specifies the precision of the model:
\itemize{
\item \code{"int8"}: 8-bit integer precision, which is the most efficient in terms of speed and memory usage.
\item \code{"fp16"}: 16-bit floating point precision, which is a good balance between speed and accuracy.
\item \code{"fp32"}: 32-bit floating point precision, which is the most accurate but also the slowest and most memory-intensive.
Protobuf and geographic models only support \code{"fp32"} precision.
}

To get an overview of all valid model option combinations use \code{\link[=model_options]{model_options()}}.
}
\examples{
# Load a pre-trained acoustic model
\dontrun{
model  <- load_model(type = "acoustic", version = "2.4", backend = "tf", precision = "int8")
}
}
