% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{new_birdnet_model}
\alias{new_birdnet_model}
\title{Create a new BirdNET model object}
\usage{
new_birdnet_model(x, ..., subclass = character())
}
\arguments{
\item{x}{A Python object representing the BirdNET model. This is typically a Python model
object created using the \code{reticulate} package.}

\item{...}{Additional attributes to attach to the BirdNET model object.}

\item{subclass}{Character. An optional subclass name for the BirdNET model (e.g., "tflite_v2.4").
The subclass is combined with the base class \code{birdnet_model}.}
}
\value{
An S3 object of class \code{birdnet_model} (and any specified subclass) containing the Python model object
and any additional attributes passed in \code{...}.
}
\description{
This function creates a new BirdNET model object by wrapping a Python model object and assigning
it a class and optional subclass. The model is created as an R object that can be interacted with
using R's S3 method dispatch.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
py_birdnet_models <- reticulate::import("birdnet.models")
tflite_model <- py_birdnet_models$v2m4$AudioModelV2M4TFLite()
birdnet_model <- new_birdnet_model(tflite_model, language = "en_us", version = "v2.4")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
