% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.birdnet_model_acoustic}
\alias{predict.birdnet_model_acoustic}
\title{Predict species from audio files using a BirdNET acoustic model}
\usage{
\method{predict}{birdnet_model_acoustic}(
  model,
  files,
  min_confidence = 0.1,
  min_confidence_custom = NULL,
  top_k = 5L,
  overlap = 0,
  apply_sigmoid = TRUE,
  sigmoid_sensitivity = 1,
  bandpass_fmin = 0L,
  bandpass_fmax = 15000L,
  species_list = NULL,
  progress = c("minimal", "progress", "benchmark")
)
}
\arguments{
\item{model}{A BirdNET model object of class \code{birdnet_model_acoustic} created with \link{load_model(type = "acoustic")}.}

\item{files}{A character vector of file paths to audio files.}

\item{min_confidence}{A numeric value to set the minimum confidence threshold for predictions.}

\item{min_confidence_custom}{A named list where each element is a single numeric value to set custom minimum confidence thresholds for specific species. A custom threshold will override the default one.}

\item{top_k}{An integer specifying the number of top predictions to return for each time interval if above minimum confidence threshold.}

\item{overlap}{A numeric value specifying the overlap duration in seconds between consecutive time intervals. Must be in the interval [0.0, 3.0].}

\item{apply_sigmoid}{A logical value indicating whether to apply a sigmoid function to the confidence scores.}

\item{sigmoid_sensitivity}{A numeric value that adjusts the sensitivity of the sigmoid function.  Must be in the interval [0.5, 1.5].}

\item{bandpass_fmin, bandpass_fmax}{A integer value to set minimum and maximum frequencies for the bandpass filter (in Hz).}

\item{species_list}{A character vector or list of species names to filter the predictions. If \code{NULL}, all species are considered.}

\item{progress}{A character string specifying the type of progress reporting. Options are "minimal", "progress", or "benchmark".}
}
\value{
An S3 object of class \code{birdnet_prediction_acoustic} and \code{birdnet_prediction} containing the prediction results.
}
\description{
This function predicts species from audio files using a BirdNET acoustic model.
}
\details{
#' @details
\subsection{Sigmoid Activation}{

When \code{apply_sigmoid = TRUE}, the raw logit scores from the linear classifier are passed
through a sigmoid function, scaling them into the range [0, 1]. This unitless confidence
score reflects BirdNET’s certainty in its prediction (it is not a direct probability of species presence).
Adjusting the \code{sigmoid_sensitivity} parameter modifies the score distribution:
\itemize{
\item Values \strong{< 1} tend to produce more extreme scores (closer to 0 or 1).
\item Values \strong{> 1} result in scores that are more moderate (centered around intermediate values).
}

For additional details on BirdNET confidence scores and guidelines for converting them to probabilities, see Wood & Kahl (2024).
}
}
\examples{
\dontrun{
# Load a BirdNET acoustic model
model <- load_model(type = "acoustic")
# Predict species from audio files
audio_file <- system.file("extdata", "soundscape.mp3", package = "birdnetR")
predictions <- predict(model, files = audio_file)
# convert predictions to a data frame
as.data.frame(predictions)
}

}
\references{
Wood, C. M., & Kahl, S. (2024). Guidelines for appropriate use of BirdNET scores and other detector outputs. Journal of Ornithology. https://doi.org/10.1007/s10336-024-02144-5
}
