% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.birdnet_model_geo}
\alias{predict.birdnet_model_geo}
\title{Predict species using a BirdNET geo model}
\usage{
\method{predict}{birdnet_model_geo}(object, latitude, longitude, week = NULL, min_confidence = 0.1, ...)
}
\arguments{
\item{object}{A BirdNET model object of class \code{birdnet_model_geo} created with \code{\link[=load_model]{load_model()}}.}

\item{latitude}{A numeric value representing the latitude of the location.}

\item{longitude}{A numeric value representing the longitude of the location.}

\item{week}{An integer value representing the week of the year (1-52).}

\item{min_confidence}{A numeric value to set the minimum confidence threshold for predictions.}

\item{...}{Additional arguments passed to the generic (currently unused).}
}
\value{
An S3 object of class \code{birdnet_prediction_geo} and \code{birdnet_prediction} containing the prediction results.
}
\description{
This function predicts species occurence for a location and week of the year using a BirdNET geo model.
}
\examples{
\dontrun{
# Load a BirdNET geo model
model <- load_model(type = "geo")
# Predict species for a specific location and week
predictions <- predict(model, latitude = 50.8334, longitude = 12.9231, week = 18L)
# Convert predictions to a data frame
as.data.frame(predictions)
}

}
