% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{predict_species}
\alias{predict_species}
\title{Predict Species Within an Audio File}
\usage{
predict_species(
  model,
  audio_path = system.file("extdata", "soundscape.wav", package = "birdnet")
)
}
\arguments{
\item{model}{An instance of the BirdNET model.}

\item{audio_path}{The path to the audio file.}
}
\value{
A list of predictions.
}
\description{
This function predicts species within an audio file using the BirdNET model.
}
