% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{predict_species}
\alias{predict_species}
\title{Predict Species Within an Audio File}
\usage{
predict_species(
  model,
  audio_file = system.file("extdata", "soundscape.wav", package = "birdnetR"),
  min_confidence = 0.1,
  batch_size = 1L,
  use_bandpass = TRUE,
  bandpass_fmin = 0L,
  bandpass_fmax = 15000L,
  apply_sigmoid = TRUE,
  sigmoid_sensitivity = 1,
  filter_species = NULL,
  file_splitting_duration_s = 600,
  keep_empty = TRUE,
  ...
)
}
\arguments{
\item{model}{BirdNETModel. An instance of the BirdNET model returned by \code{\link[=init_model]{init_model()}}.}

\item{audio_file}{character. The path to the audio file.}

\item{min_confidence}{numeric. Minimum confidence threshold for predictions.}

\item{batch_size}{integer. Number of audio samples to process in a batch.}

\item{use_bandpass}{logical. Whether to apply a bandpass filter.}

\item{bandpass_fmin, bandpass_fmax}{numeric. Minimum/Maximumfrequency for the bandpass filter (in Hz). Ignored if \code{use_bandpass} is False.}

\item{apply_sigmoid}{logical. Whether to apply a sigmoid function to the model output.}

\item{sigmoid_sensitivity}{numeric. Sensitivity parameter for the sigmoid function. Must be in the interval 0.5 - 1.5. Ignored if \code{apply_sigmoid} is False.}

\item{filter_species}{character or NULL. A set of species to filter the predictions. If NULL, no filtering is applied.}

\item{file_splitting_duration_s}{numeric. Duration in seconds for splitting the audio file into smaller segments for processing.}

\item{keep_empty}{logical. Whether to include empty intervals in the output.}

\item{...}{Additional arguments passed to the prediction function.}
}
\value{
A data frame with columns: \code{start}, \code{end}, \code{scientific_name}, \code{common_name}, and \code{confidence}.
Each row represents a single prediction.
}
\description{
This function predicts species within an audio file using the BirdNET model.
}
