% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_interface.R
\name{predict_species_from_audio_file}
\alias{predict_species_from_audio_file}
\alias{predict_species_from_audio_file.birdnet_model}
\title{Predict species within an audio file using a BirdNET model}
\usage{
predict_species_from_audio_file(
  model,
  audio_file,
  min_confidence = 0.1,
  batch_size = 1L,
  chunk_overlap_s = 0,
  use_bandpass = TRUE,
  bandpass_fmin = 0L,
  bandpass_fmax = 15000L,
  apply_sigmoid = TRUE,
  sigmoid_sensitivity = 1,
  filter_species = NULL,
  keep_empty = TRUE
)

\method{predict_species_from_audio_file}{birdnet_model}(
  model,
  audio_file,
  min_confidence = 0.1,
  batch_size = 1L,
  chunk_overlap_s = 0,
  use_bandpass = TRUE,
  bandpass_fmin = 0L,
  bandpass_fmax = 15000L,
  apply_sigmoid = TRUE,
  sigmoid_sensitivity = 1,
  filter_species = NULL,
  keep_empty = TRUE
)
}
\arguments{
\item{model}{A BirdNET model object. An instance of the BirdNET model (e.g., \code{birdnet_model_tflite}, \code{birdnet_model_protobuf}).}

\item{audio_file}{character. The path to the audio file.}

\item{min_confidence}{numeric. Minimum confidence threshold for predictions (default is 0.1).}

\item{batch_size}{integer. Number of audio samples to process in a batch (default is 1L).}

\item{chunk_overlap_s}{numeric. The overlap between audio chunks in seconds (default is 0). Must be in the interval [0.0, 3.0].}

\item{use_bandpass}{logical. Whether to apply a bandpass filter (default is TRUE).}

\item{bandpass_fmin, bandpass_fmax}{integer. Minimum and maximum frequencies for the bandpass filter (in Hz). Ignored if \code{use_bandpass} is FALSE (default is 0L to 15000L).}

\item{apply_sigmoid}{logical. Whether to apply a sigmoid function to the model output (default is TRUE).}

\item{sigmoid_sensitivity}{numeric. Sensitivity parameter for the sigmoid function (default is 1). Must be in the interval [0.5, 1.5]. Ignored if \code{apply_sigmoid} is FALSE.}

\item{filter_species}{NULL, a character vector of length greater than 0, or a list where each element is a single non-empty character string. Used to filter the predictions. If NULL (default), no filtering is applied.}

\item{keep_empty}{logical. Whether to include empty intervals in the output (default is TRUE).}
}
\value{
A data frame with columns: \code{start}, \code{end}, \code{scientific_name}, \code{common_name}, and \code{confidence}. Each row represents a single prediction.
}
\description{
Use a BirdNET model to predict species within an audio file. The model can be a TFLite model, a custom model, or a Protobuf model.
}
\details{
Applying a sigmoid activation function (\code{apply_sigmoid=TRUE}) scales the unbound class output of the linear classifier ("logit score") to the range \code{0-1}.
This confidence score is a unitless, numeric expression of BirdNET’s “confidence” in its prediction (but not the probability of species presence).
Sigmoid sensitivity < 1 leads to more higher and lower scoring predictions, and a value > 1 leads to more intermediate-scoring predictions.

For more information on BirdNET confidence scores, the sigmoid activation function, and a suggested workflow on how to convert confidence scores to probabilities, see Wood & Kahl, 2024.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(birdnetR)

model <- birdnet_model_tflite(version = "v2.4", language = "en_us")
predictions <- predict_species_from_audio_file(model, "path/to/audio.wav", min_confidence = 0.2)
\dontshow{\}) # examplesIf}
}
\references{
Wood, C. M., & Kahl, S. (2024). Guidelines for appropriate use of BirdNET scores and other detector outputs. Journal of Ornithology. https://doi.org/10.1007/s10336-024-02144-5
}
\seealso{
\code{\link[=get_species_from_file]{get_species_from_file()}} for more details on species filtering.

\code{\link{predict_species_from_audio_file.birdnet_model}}
}
